good baseline :-(, I'm doing something slightly wrong in downloading
The goal is to reliably get pages out into a file.  It would help to
have a test that only does this thing, anyway, and then instrument the
thing to try many strategies.  Maybe the right download strategy will
become obvious with a tool designed to explicitly expose the problems
with a given download strategy, and given a way to exercise it many
times.

So, these tools are insufficient to answer the remaining questions:

  * if downloading is done correctly, do the radio sent/receive stats
    stay on par?

  * how do we correctly aggregate data across many reads?
    * if we encounter a CRC mismatch, how do we respond?
      * there is data left on the rf to read; failure to read it
        results in errors, radio sequence and timeouts, without some
        recovery strategy to clear the buffer
      * when greedily attempting to clear the RF buffer, eg blindly
        retrying, using only sizes from poll_size to format the
        ReadRadio packets, there is unexpected amounts of data in the
        buffer.
        * when doing this, there is a mismatch in radio receive/sent
          counters, with a slowly incrementing timeout