add discussion of special bolus
I think this and the midnight records are the only ones left.
As for these special bolus opcodes, 0x5c, I suspect that 0x5b is it's own
opcode, ~11 body, and that 0x01 is it's own opcode (analysis shows it
always parses cleanly), and 0x5c is also it's own opcode.

0x5x seems to always require special parsing, always read up until the
null byte?  The open question is whether or not there are two null
bytes, and how to know if that's the case?