allow some breathing room between fast polling loops
When sustaining many dozens/hundreds of commands all at once, the pump
sometimes needs a little breathing room.

According to the analysis, there should be some byte which describes
the need for the comms to pause for some time, ~3 seconds, and/or
until an ACK is recieved.  Instead, sometimes introduce longer delays
before attempting to continue, if it seems like communications have
hung.