cgm download pages identical
When downloading the CGM data from the MM pump. The download command
seems to be downloading the same page over and over again. All the pages
downloaded (1-9) appear to be identical.

Could be that the data in them is identical and that the downloading
process is working correctly, but it just seems a little odd.