bolus: force explicit selection of stroke size
Make the bolus tool safer; the stroke sizes for pumps vary widely by model.
This script has only been tested on 515.

This change forces the user to explicitly declare either the pump
model or the stroke size to use.  This should allow easily re-using
the tool to discover stroke sizes fairly easily, in order to add the
pump specific option.

The tool will error out and refuse to do anything if the
strokes_per_unit has not been explicitly indicated, either through
specifying the pump model or the strokes option.