Attempt BolusWizard MMOL support, based on model
For @sulkaharo and others.
This is a quick hack, see:
* https://gist.github.com/sulkaharo/3281f729b7c52e10c69c

* `bg`
* `bg_target_low`
* `bg_target_high`
* `sensitivity`

For these fields, divide by ten if the model is set to MMOL_DEFAULT.
AFAIK, these models are not sold in the US and probably have a European profile
by default.

However, it is highly likely that these models can be run with US profiles, and
I've seen both European and US 522 pumps, however, there's not a way for
end-users to reset the Locale or units.  The Wizard settings do indicate the
units in the payload, and match the Wizard's records here unless there's been
some change unaccounted for (analyzing stale data engineering data in house,
maybe?).